add get date component
added get upgrade date helper so it can be used in multiple places. Also specified upgrade page fbt descriptions
